+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 The following is a curated list of changes in the Enact project, newest changes on the top.
 
+## [5.5.1] - 2026-06-01
+
+### Changed
+
+- `ui/VirtualList.VirtualListBasic` scroll animation method for `scrollMode: native`
+
+### Fixed
+
+- `ui/Scroller` jumping back when scrolled by long press
+- `ui/VirtualList` overscroll effect at the bottom of the list
+- `ui/VirtualList.VirtualList` overscroll effect at the bottom of the list
+- `ui/VirtualList.VirtualGridList` overscroll effect at the bottom of the list
+
+## [5.4.4] - 2026-05-22
+
+### Fixed
+
+- `ui/Scroller` jumping back when scrolled by long press
+
 ## [5.5.0] - 2026-05-08
 
 ### Added